Ignores unwanted files
Here we are with a fresh repository. Check out the size! Its tiny!

Let's try to keep it small (even for a monorepo).

This ignores commonly unwanted files (no zips or tarballs please) and
will surely be updated as the project grows.

But let's all try to keep it small :D